+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.1] - 2024-05-20
 
 ### Fixed
 
@@ -110,7 +110,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial version of web app
 
-[Unreleased]: https://github.com/treiher/valens/compare/v0.4.0...HEAD
+[0.4.1]: https://github.com/treiher/valens/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/treiher/valens/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/treiher/valens/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/treiher/valens/compare/v0.1.0...v0.2.0